Improved UX, added error popup messages, normalized venv

+++ lol_manager.py
@@ -1,5 +1,5 @@
 import argparse
-import tkinter
+import psutil
 from enum import Enum
 from time import sleep
 from tkinter import Tk
@@ -7,11 +7,16 @@ from tkinter.simpledialog import askstring
 from pykeepass import PyKeePass
 from pykeepass.exceptions import CredentialsError
 from pynput.keyboard import Key, Controller
-import psutil as psutil
 from pywinauto import Application, Desktop
 from sources.accounts import ACCOUNTS, Account
+from sources.error_popup import error_popup
 from sources.password_popup import get_password
 from sources.settings import SETTINGS
+
+
+class InvalidSettings(Exception):
+    def __init__(self, message):
+        super().__init__(message)
 
 
 class InvalidAccount(Exception):
@@ -56,6 +61,9 @@ def get_credentials(account: Account, behavior: LoginBehavior = LoginBehavior.US
 
     print(f'get_credentials: use_keepass = {use_keepass}')
     if use_keepass:
+        if not SETTINGS.keepass_path:
+            raise InvalidCredentials('Keepass path not set')
+
         master_key = get_password()
         try:
             keepass = PyKeePass(SETTINGS.keepass_path, password=master_key)
@@ -74,13 +82,26 @@ def get_credentials(account: Account, behavior: LoginBehavior = LoginBehavior.US
 
 
 def stop_lol_client():
+    if not SETTINGS.lol_path:
+        raise InvalidSettings('League of Legends path not set')
+
+    if not SETTINGS.lol_process_names:
+        raise InvalidSettings('League of Legends process names not set')
+
     for process in psutil.process_iter():
         if process.name() in SETTINGS.lol_process_names:
             process.terminate()
 
 
 def start_lol_client() -> Desktop:
-    Application(backend='uia').start(SETTINGS.lol_client_path + ' ' + SETTINGS.lol_client_args)
+    if not SETTINGS.lol_path:
+        raise InvalidSettings('League of Legends path not set')
+
+    try:
+        Application(backend='uia').start(SETTINGS.lol_client_path + ' ' + SETTINGS.lol_client_args)
+    except Exception:
+        raise InvalidSettings(f'Invalid League of Legends path')
+
     return Desktop(backend='uia').window(title_re='.*Riot Client Main.*')
 
 
@@ -91,17 +112,26 @@ def restart_lol_client():
 
 def login_lol_client(name: str, behavior: LoginBehavior = LoginBehavior.USE_SETTINGS):
     print(f'Logging in {name}, behavior = {behavior}')
+
     try:
         account = get_account(name)
     except InvalidAccount as e:
         print(e)
+        error_popup(message=str(e))
         return
     try:
         username, password = get_credentials(account, behavior)
     except (InvalidAccount, InvalidCredentials) as e:
         print(e)
+        error_popup(message=str(e))
         return
-    window = restart_lol_client()
+
+    try:
+        window = restart_lol_client()
+    except InvalidSettings as e:
+        print(e)
+        error_popup(message=str(e))
+        return
 
     for i in range(20):
         if window.exists():
@@ -110,7 +140,9 @@ def login_lol_client(name: str, behavior: LoginBehavior = LoginBehavior.USE_SETT
         sleep(0.1)
 
     if not window.exists():
-        print('Riot Client Main window not found')
+        message = 'Riot Client Main window not found'
+        print(message)
+        error_popup(message=message)
         return
 
     window.set_focus()
@@ -127,10 +159,13 @@ def login_lol_client(name: str, behavior: LoginBehavior = LoginBehavior.USE_SETT
     keyboard.tap(Key.enter)
 
 
-if __name__ == "__main__":
+def main():
     parser = argparse.ArgumentParser()
     parser.add_argument('--name', required=True, help='The name of the account to login')
     parser.add_argument('--behavior', required=False, help='Valid values = [use_keepass, use_settings, never_use_keepass]')
     args = parser.parse_args()
     login_lol_client(name=str(args.name), behavior=LoginBehavior(args.behavior))
 
+
+if __name__ == "__main__":
+    main()
